Refactor the files and removed the duplicates
